Fix a BinderServerTransport crash in the rare shutdown-before-start case (#12440)
The `isShutdown()` clause of `BinderServerTransport#start()` code was completely
untested and did not in fact work.

The problem is that if the listener does in fact arrive via start()
after shutdown, BinderTransport's `shutdownInternal()` has already set
the state to `SHUTDOWN_TERMINATED` (which is not a valid transition from
itself). It has also already scheduled a call to notifyTerminated() and
releaseExecutors(). This causes a duplicate call to
`transportTerminated` and releasing the same executor twice.

This commit changes `start()` to leave changing state and releasing
executors to `shutdownInternal()`'s. `notifyTerminated()` either runs
then (if already started) or from within `start()` (if not yet started)

Fixes #12439.