binder: Move internal test utils to src/test/
testFixtures is for things depending on binder to use for their own
tests. Classes that are only used by binder tests should just live in
src/test/.

This was noticed because FakeDeadBinder was in testFixtures _and not_ in
internal, even though it clearly shouldn't be used outside of binder.
FakeDeadBinder was moved to the internal package to better mirror its
usage.