interop-testing: timeoutOnSleepingServer should fail on exception
This is ancient code, but our API should not throw IllegalStateException
in the case of races. This essentially reverts 0958fd407. The stream
code has changed a lot since this code was introduced, but at the very
least the replacement of AbstractStream that culminated in b661ac7d7
means the bug is almost certainly gone.

I ran Http2Test 1000 times and there were no failures.