compiler: Upgrade Bazel protobuf to 33.1 (#12553)
Following up on
https://github.com/grpc/grpc-java/commit/58ae5f808cf8e20c5864033c9a8f485b237f9dfc,
change Bazel's view of the protobuf version to 33.1 too.

This is a bit of an adventure in the bzlmod/WORKSPACE compatibility
maze. A few rulesets like `rules_java` must be upgraded too.

C++ compilation must happen until C++17 now due to the new abseil
version required.

Dependency on `rules_proto` is removed because it has become a
superfluous wrapper over rules in the upstream protobuf repo.